Move sd_manager and fix misc
